add `pubkey` and `secret_key` submodules and refactor
