--- bls/src/pod.rs
@@ -1,431 +1,0 @@
-#[cfg(feature = "bytemuck")]
-use bytemuck::{Pod, PodInOption, Zeroable, ZeroableInOption};
-use {
-    base64::{prelude::BASE64_STANDARD, Engine},
-    core::fmt,
-};
-#[cfg(feature = "serde")]
-use {
-    serde::{Deserialize, Serialize},
-    serde_with::serde_as,
-};
-
-/// Size of a BLS signature in a compressed point representation
-pub const BLS_SIGNATURE_COMPRESSED_SIZE: usize = 96;
-
-/// Size of a BLS signature in a compressed point representation in base64
-pub const BLS_SIGNATURE_COMPRESSED_BASE64_SIZE: usize = 128;
-
-/// Size of a BLS signature in an affine point representation
-pub const BLS_SIGNATURE_AFFINE_SIZE: usize = 192;
-
-/// Size of a BLS signature in an affine point representation in base64
-pub const BLS_SIGNATURE_AFFINE_BASE64_SIZE: usize = 256;
-
-/// A serialized BLS signature in a compressed point representation
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(feature = "serde", cfg_eval::cfg_eval, serde_as)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
-#[repr(transparent)]
-pub struct SignatureCompressed(
-    #[cfg_attr(feature = "serde", serde_as(as = "[_; BLS_SIGNATURE_COMPRESSED_SIZE]"))]
-    pub  [u8; BLS_SIGNATURE_COMPRESSED_SIZE],
-);
-
-impl Default for SignatureCompressed {
-    fn default() -> Self {
-        Self([0; BLS_SIGNATURE_COMPRESSED_SIZE])
-    }
-}
-
-impl fmt::Display for SignatureCompressed {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = SignatureCompressed,
-    BYTES_LEN = BLS_SIGNATURE_COMPRESSED_SIZE,
-    BASE64_LEN = BLS_SIGNATURE_COMPRESSED_BASE64_SIZE
-);
-
-/// A serialized BLS signature in an affine point representation
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(feature = "serde", cfg_eval::cfg_eval, serde_as)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
-#[repr(transparent)]
-pub struct Signature(
-    #[cfg_attr(feature = "serde", serde_as(as = "[_; BLS_SIGNATURE_AFFINE_SIZE]"))]
-    pub  [u8; BLS_SIGNATURE_AFFINE_SIZE],
-);
-
-impl Default for Signature {
-    fn default() -> Self {
-        Self([0; BLS_SIGNATURE_AFFINE_SIZE])
-    }
-}
-
-impl fmt::Display for Signature {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = Signature,
-    BYTES_LEN = BLS_SIGNATURE_AFFINE_SIZE,
-    BASE64_LEN = BLS_SIGNATURE_AFFINE_BASE64_SIZE
-);
-
-/// Size of a BLS proof of possession in a compressed point representation
-pub const BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE: usize = 96;
-
-/// Size of a BLS proof of possession in a compressed point representation in base64
-pub const BLS_PROOF_OF_POSSESSION_COMPRESSED_BASE64_SIZE: usize = 128;
-
-/// Size of a BLS proof of possession in an affine point representation
-pub const BLS_PROOF_OF_POSSESSION_AFFINE_SIZE: usize = 192;
-
-/// Size of a BLS proof of possession in an affine point representation in base64
-pub const BLS_PROOF_OF_POSSESSKON_AFFINE_BASE64_SIZE: usize = 256;
-
-/// A serialized BLS signature in a compressed point representation
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(feature = "serde", cfg_eval::cfg_eval, serde_as)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
-#[repr(transparent)]
-pub struct ProofOfPossessionCompressed(
-    #[cfg_attr(
-        feature = "serde",
-        serde_as(as = "[_; BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE]")
-    )]
-    pub [u8; BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE],
-);
-
-impl Default for ProofOfPossessionCompressed {
-    fn default() -> Self {
-        Self([0; BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE])
-    }
-}
-
-impl fmt::Display for ProofOfPossessionCompressed {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = ProofOfPossessionCompressed,
-    BYTES_LEN = BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE,
-    BASE64_LEN = BLS_PROOF_OF_POSSESSION_COMPRESSED_BASE64_SIZE
-);
-
-/// A serialized BLS signature in an affine point representation
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(feature = "serde", cfg_eval::cfg_eval, serde_as)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
-#[repr(transparent)]
-pub struct ProofOfPossession(
-    #[cfg_attr(
-        feature = "serde",
-        serde_as(as = "[_; BLS_PROOF_OF_POSSESSION_AFFINE_SIZE]")
-    )]
-    pub [u8; BLS_PROOF_OF_POSSESSION_AFFINE_SIZE],
-);
-
-impl Default for ProofOfPossession {
-    fn default() -> Self {
-        Self([0; BLS_PROOF_OF_POSSESSION_AFFINE_SIZE])
-    }
-}
-
-impl fmt::Display for ProofOfPossession {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = ProofOfPossession,
-    BYTES_LEN = BLS_PROOF_OF_POSSESSION_AFFINE_SIZE,
-    BASE64_LEN = BLS_PROOF_OF_POSSESSKON_AFFINE_BASE64_SIZE
-);
-
-/// Size of a BLS public key in a compressed point representation
-pub const BLS_PUBLIC_KEY_COMPRESSED_SIZE: usize = 48;
-
-/// Size of a BLS public key in a compressed point representation in base64
-pub const BLS_PUBLIC_KEY_COMPRESSED_BASE64_SIZE: usize = 128;
-
-/// Size of a BLS public key in an affine point representation
-pub const BLS_PUBLIC_KEY_AFFINE_SIZE: usize = 96;
-
-/// Size of a BLS public key in an affine point representation in base64
-pub const BLS_PUBLIC_KEY_AFFINE_BASE64_SIZE: usize = 256;
-
-/// A serialized BLS public key in a compressed point representation
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(feature = "serde", cfg_eval::cfg_eval, serde_as)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-#[repr(transparent)]
-pub struct PubkeyCompressed(
-    #[cfg_attr(
-        feature = "serde",
-        serde_as(as = "[_; BLS_PUBLIC_KEY_COMPRESSED_SIZE]")
-    )]
-    pub [u8; BLS_PUBLIC_KEY_COMPRESSED_SIZE],
-);
-
-impl Default for PubkeyCompressed {
-    fn default() -> Self {
-        Self([0; BLS_PUBLIC_KEY_COMPRESSED_SIZE])
-    }
-}
-
-impl fmt::Display for PubkeyCompressed {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = PubkeyCompressed,
-    BYTES_LEN = BLS_PUBLIC_KEY_COMPRESSED_SIZE,
-    BASE64_LEN = BLS_PUBLIC_KEY_COMPRESSED_BASE64_SIZE
-);
-
-/// A serialized BLS public key in an affine point representation
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(feature = "serde", cfg_eval::cfg_eval, serde_as)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, Eq, Ord, PartialEq, PartialOrd)]
-#[repr(transparent)]
-pub struct Pubkey(
-    #[cfg_attr(feature = "serde", serde_as(as = "[_; BLS_PUBLIC_KEY_AFFINE_SIZE]"))]
-    pub  [u8; BLS_PUBLIC_KEY_AFFINE_SIZE],
-);
-
-impl Default for Pubkey {
-    fn default() -> Self {
-        Self([0; BLS_PUBLIC_KEY_AFFINE_SIZE])
-    }
-}
-
-impl fmt::Display for Pubkey {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = Pubkey,
-    BYTES_LEN = BLS_PUBLIC_KEY_AFFINE_SIZE,
-    BASE64_LEN = BLS_PUBLIC_KEY_AFFINE_BASE64_SIZE
-);
-
-// Byte arrays are both `Pod` and `Zeraoble`, but the traits `bytemuck::Pod` and
-// `bytemuck::Zeroable` can only be derived for power-of-two length byte arrays.
-// Directly implement these traits for types that are simple wrappers around
-// byte arrays.
-#[cfg(feature = "bytemuck")]
-mod bytemuck_impls {
-    use super::*;
-    unsafe impl Zeroable for PubkeyCompressed {}
-    unsafe impl Pod for PubkeyCompressed {}
-    unsafe impl ZeroableInOption for PubkeyCompressed {}
-    unsafe impl PodInOption for PubkeyCompressed {}
-
-    unsafe impl Zeroable for Pubkey {}
-    unsafe impl Pod for Pubkey {}
-    unsafe impl ZeroableInOption for Pubkey {}
-    unsafe impl PodInOption for Pubkey {}
-
-    unsafe impl Zeroable for Signature {}
-    unsafe impl Pod for Signature {}
-    unsafe impl ZeroableInOption for Signature {}
-    unsafe impl PodInOption for Signature {}
-
-    unsafe impl Zeroable for SignatureCompressed {}
-    unsafe impl Pod for SignatureCompressed {}
-    unsafe impl ZeroableInOption for SignatureCompressed {}
-    unsafe impl PodInOption for SignatureCompressed {}
-
-    unsafe impl Zeroable for ProofOfPossessionCompressed {}
-    unsafe impl Pod for ProofOfPossessionCompressed {}
-    unsafe impl ZeroableInOption for ProofOfPossessionCompressed {}
-    unsafe impl PodInOption for ProofOfPossessionCompressed {}
-
-    unsafe impl Zeroable for ProofOfPossession {}
-    unsafe impl Pod for ProofOfPossession {}
-    unsafe impl ZeroableInOption for ProofOfPossession {}
-    unsafe impl PodInOption for ProofOfPossession {}
-}
-
-macro_rules! impl_from_str {
-    (TYPE = $type:ident, BYTES_LEN = $bytes_len:expr, BASE64_LEN = $base64_len:expr) => {
-        impl core::str::FromStr for $type {
-            type Err = crate::error::BlsError;
-
-            fn from_str(s: &str) -> Result<Self, Self::Err> {
-                use base64::Engine;
-
-                if s.len() > $base64_len {
-                    return Err(Self::Err::ParseFromString);
-                }
-                let mut bytes = [0u8; $bytes_len];
-                let decoded_len = base64::prelude::BASE64_STANDARD
-                    .decode_slice(s, &mut bytes)
-                    .map_err(|_| Self::Err::ParseFromString)?;
-                if decoded_len != $bytes_len {
-                    Err(Self::Err::ParseFromString)
-                } else {
-                    Ok($type(bytes))
-                }
-            }
-        }
-    };
-}
-use impl_from_str;
-
-#[cfg(test)]
-mod tests {
-    use {super::*, crate::keypair::Keypair, core::str::FromStr, std::string::ToString};
-
-    #[test]
-    fn pubkey_from_str() {
-        let pubkey = Keypair::new().public;
-        let pubkey_affine: Pubkey = pubkey.into();
-        let pubkey_affine_string = pubkey_affine.to_string();
-        let pubkey_affine_from_string = Pubkey::from_str(&pubkey_affine_string).unwrap();
-        assert_eq!(pubkey_affine, pubkey_affine_from_string);
-
-        let pubkey_compressed = PubkeyCompressed([1; BLS_PUBLIC_KEY_COMPRESSED_SIZE]);
-        let pubkey_compressed_string = pubkey_compressed.to_string();
-        let pubkey_compressed_from_string =
-            PubkeyCompressed::from_str(&pubkey_compressed_string).unwrap();
-        assert_eq!(pubkey_compressed, pubkey_compressed_from_string);
-    }
-
-    #[test]
-    fn signature_from_str() {
-        let signature_affine = Signature([1; BLS_SIGNATURE_AFFINE_SIZE]);
-        let signature_affine_string = signature_affine.to_string();
-        let signature_affine_from_string = Signature::from_str(&signature_affine_string).unwrap();
-        assert_eq!(signature_affine, signature_affine_from_string);
-
-        let signature_compressed = SignatureCompressed([1; BLS_SIGNATURE_COMPRESSED_SIZE]);
-        let signature_compressed_string = signature_compressed.to_string();
-        let signature_compressed_from_string =
-            SignatureCompressed::from_str(&signature_compressed_string).unwrap();
-        assert_eq!(signature_compressed, signature_compressed_from_string);
-    }
-
-    #[test]
-    fn proof_of_possession_from_str() {
-        let proof_of_possession = ProofOfPossession([1; BLS_PROOF_OF_POSSESSION_AFFINE_SIZE]);
-        let proof_of_possession_string = proof_of_possession.to_string();
-        let proof_of_possession_from_string =
-            ProofOfPossession::from_str(&proof_of_possession_string).unwrap();
-        assert_eq!(proof_of_possession, proof_of_possession_from_string);
-
-        let proof_of_possession_compressed =
-            ProofOfPossessionCompressed([1; BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE]);
-        let proof_of_possession_compressed_string = proof_of_possession_compressed.to_string();
-        let proof_of_possession_compressed_from_string =
-            ProofOfPossessionCompressed::from_str(&proof_of_possession_compressed_string).unwrap();
-        assert_eq!(
-            proof_of_possession_compressed,
-            proof_of_possession_compressed_from_string
-        );
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn serialize_and_deserialize_pubkey() {
-        let original = Pubkey::default();
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: Pubkey = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-
-        let original = Pubkey([1; BLS_PUBLIC_KEY_AFFINE_SIZE]);
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: Pubkey = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn serialize_and_deserialize_pubkey_compressed() {
-        let original = PubkeyCompressed::default();
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: PubkeyCompressed = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-
-        let original = PubkeyCompressed([1; BLS_PUBLIC_KEY_COMPRESSED_SIZE]);
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: PubkeyCompressed = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn serialize_and_deserialize_signature() {
-        let original = Signature::default();
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: Signature = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-
-        let original = Signature([1; BLS_SIGNATURE_AFFINE_SIZE]);
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: Signature = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn serialize_and_deserialize_signature_compressed() {
-        let original = SignatureCompressed::default();
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: SignatureCompressed = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-
-        let original = SignatureCompressed([1; BLS_SIGNATURE_COMPRESSED_SIZE]);
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: SignatureCompressed = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn serialize_and_deserialize_proof_of_possession() {
-        let original = ProofOfPossession::default();
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: ProofOfPossession = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-
-        let original = ProofOfPossession([1; BLS_PROOF_OF_POSSESSION_AFFINE_SIZE]);
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: ProofOfPossession = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn serialize_and_deserialize_proof_of_possession_compressed() {
-        let original = ProofOfPossessionCompressed::default();
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: ProofOfPossessionCompressed = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-
-        let original = ProofOfPossessionCompressed([1; BLS_PROOF_OF_POSSESSION_COMPRESSED_SIZE]);
-        let serialized = bincode::serialize(&original).unwrap();
-        let deserialized: ProofOfPossessionCompressed = bincode::deserialize(&serialized).unwrap();
-        assert_eq!(original, deserialized);
-    }
-}
